image 2/requirements.txt
flask==3.0.2
python-docx==1.1.0
werkzeug==3.0.1
watchdog==3.0.0 